% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_tabulate_by_year.R
\name{cross_tabulate_by_year}
\alias{cross_tabulate_by_year}
\title{This function allows you to easily create crosstabs for cross-sectional analysis.
Note: This function is still in a beta phase. Future updates should include customizable rouding, total rows, and updated syntax.}
\usage{
cross_tabulate_by_year(data, weight, y_col, x_col, year_col)
}
\description{
This function allows you to easily create crosstabs for cross-sectional analysis.
Note: This function is still in a beta phase. Future updates should include customizable rouding, total rows, and updated syntax.
}
