% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unweighted crosstab.R
\name{unweighted_crosstab}
\alias{unweighted_crosstab}
\title{normal crosstab without survey weights.
x_order allows you to quickly arrange the order of variables in the x axis. This is particularlly useful to prepare dataframes for table creation.}
\usage{
unweighted_crosstab(data, x, y, digits, x_order = NULL)
}
\description{
normal crosstab without survey weights.
x_order allows you to quickly arrange the order of variables in the x axis. This is particularlly useful to prepare dataframes for table creation.
}
