% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted average.R
\name{weighted_average}
\alias{weighted_average}
\title{This function allows users to calculate means while incorportating survey weights.
The function differs from weighted_freq because it doesn't create a frequency table. Instead, it produces a mean.
This function was developed by Alexander Bills.}
\usage{
weighted_average(data, var, group, weight)
}
\description{
This function allows users to calculate means while incorportating survey weights.
The function differs from weighted_freq because it doesn't create a frequency table. Instead, it produces a mean.
This function was developed by Alexander Bills.
}
