% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive function.R, R/weighted crosstab.R
\name{weighted_crosstab}
\alias{weighted_crosstab}
\title{This function allows users to simultaneously run weighted crosstabs.
Users input a master dataframe that contains all the specifications for each weighted crosstab.
In the master dataframe, users specify the source data the crosstab should draw from. The source data MUST be in the global enviroment.
Other columns include necesasrry specifications for weighted crosstabs: x variable, y variable, and weights.
Users can also specify any necesarry filtering in the master dataframe.
Essentially, the specifications for each crosstab are configured in the master dataframe, generally via Excel.}
\usage{
weighted_crosstab(data, weight, y, x, digits = 1, total_row = FALSE)

weighted_crosstab(data, weight, y, x, digits = 1, total_row = FALSE)
}
\description{
This function allows users to simultaneously run weighted crosstabs.
Users input a master dataframe that contains all the specifications for each weighted crosstab.
In the master dataframe, users specify the source data the crosstab should draw from. The source data MUST be in the global enviroment.
Other columns include necesasrry specifications for weighted crosstabs: x variable, y variable, and weights.
Users can also specify any necesarry filtering in the master dataframe.
Essentially, the specifications for each crosstab are configured in the master dataframe, generally via Excel.

This is the key function of the CougarRA package. Weighted crosstab allows users to easily created and customize weighted crosstabs.
The function improves on the retired cross_tabulate function in several ways.
First, weighted crosstab allows you to customize rounding. The default rounding is set to the tenths place, but that can easily be adjusted using the digits parameter within the function.
Second, weighted crosstab allows users to choose whether to include a total row for the N of each value of x. The default setting is no total column, but that can be adjusted by setting total_column to TRUE.
Third, weighted crosstab has improved efficiency. Error messages are more specific, allowing users to easily troubleshoot. Likewise, weighted crosstab uses updated syntax, eliminating error messages for depreciated codes.
}
