% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_freq_table.R
\name{weighted_freq}
\alias{weighted_freq}
\title{This function creates a weighted frequency table.}
\usage{
weighted_freq(data, weight, variable)
}
\description{
This function creates a weighted frequency table.
}
