% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_frequency_by_year.R
\name{weighted_frequency_by_year}
\alias{weighted_frequency_by_year}
\title{This function allows users to create a simple, one variable frequency table for cross-sectional analysis.
The user simply states the variable in the call, the weight, and the year column}
\usage{
weighted_frequency_by_year(data, weight, variable, year_col)
}
\description{
This function allows users to create a simple, one variable frequency table for cross-sectional analysis.
The user simply states the variable in the call, the weight, and the year column
}
